Move `specific` and related methods into a mixin
