+++ wagtail/models/__init__.py
@@ -31,7 +31,7 @@ from django.core.handlers.base import BaseHandler
 from django.core.handlers.wsgi import WSGIRequest
 from django.core.serializers.json import DjangoJSONEncoder
 from django.db import models, transaction
-from django.db.models import DEFERRED, Q, Value
+from django.db.models import Q, Value
 from django.db.models.expressions import OuterRef, Subquery
 from django.db.models.functions import Cast, Concat, Substr
 from django.dispatch import receiver
@@ -120,6 +120,7 @@ from .i18n import (  # noqa
 )
 from .reference_index import ReferenceIndex  # noqa
 from .sites import Site, SiteManager, SiteRootPath  # noqa
+from .specific import SpecificMixin
 from .view_restrictions import BaseViewRestriction
 
 logger = logging.getLogger("wagtail")
@@ -1051,6 +1052,7 @@ class AbstractPage(
     LockableMixin,
     RevisionMixin,
     TranslatableMixin,
+    SpecificMixin,
     MP_Node,
 ):
     """
@@ -1522,123 +1524,6 @@ class Page(AbstractPage, index.Indexed, ClusterableModel, metaclass=PageBase):
                 )
             )
         )
-
-    def get_specific(self, deferred=False, copy_attrs=None, copy_attrs_exclude=None):
-        """
-        Return this page in its most specific subclassed form.
-
-        By default, a database query is made to fetch all field values for the
-        specific object. If you only require access to custom methods or other
-        non-field attributes on the specific object, you can use
-        ``deferred=True`` to avoid this query. However, any attempts to access
-        specific field values from the returned object will trigger additional
-        database queries.
-
-        By default, references to all non-field attribute values are copied
-        from current object to the returned one. This includes:
-
-        * Values set by a queryset, for example: annotations, or values set as
-          a result of using ``select_related()`` or ``prefetch_related()``.
-        * Any ``cached_property`` values that have been evaluated.
-        * Attributes set elsewhere in Python code.
-
-        For fine-grained control over which non-field values are copied to the
-        returned object, you can use ``copy_attrs`` to specify a complete list
-        of attribute names to include. Alternatively, you can use
-        ``copy_attrs_exclude`` to specify a list of attribute names to exclude.
-
-        If called on a page object that is already an instance of the most
-        specific class (e.g. an ``EventPage``), the object will be returned
-        as is, and no database queries or other operations will be triggered.
-
-        If the page was originally created using a page type that has since
-        been removed from the codebase, a generic ``Page`` object will be
-        returned (without any custom field values or other functionality
-        present on the original class). Usually, deleting these pages is the
-        best course of action, but there is currently no safe way for Wagtail
-        to do that at migration time.
-        """
-        model_class = self.specific_class
-
-        if model_class is None:
-            # The codebase and database are out of sync (e.g. the model exists
-            # on a different git branch and migrations were not applied or
-            # reverted before switching branches). So, the best we can do is
-            # return the page in it's current form.
-            return self
-
-        if isinstance(self, model_class):
-            # self is already an instance of the most specific class.
-            return self
-
-        if deferred:
-            # Generate a tuple of values in the order expected by __init__(),
-            # with missing values substituted with DEFERRED ()
-            values = tuple(
-                getattr(self, f.attname, self.pk if f.primary_key else DEFERRED)
-                for f in model_class._meta.concrete_fields
-            )
-            # Create object from known attribute values
-            specific_obj = model_class(*values)
-            specific_obj._state.adding = self._state.adding
-        else:
-            # Fetch object from database
-            specific_obj = model_class._default_manager.get(id=self.id)
-
-        # Copy non-field attribute values
-        if copy_attrs is not None:
-            for attr in (attr for attr in copy_attrs if attr in self.__dict__):
-                setattr(specific_obj, attr, getattr(self, attr))
-        else:
-            exclude = copy_attrs_exclude or ()
-            for k, v in ((k, v) for k, v in self.__dict__.items() if k not in exclude):
-                # only set values that haven't already been set
-                specific_obj.__dict__.setdefault(k, v)
-
-        return specific_obj
-
-    @cached_property
-    def specific(self):
-        """
-        Returns this page in its most specific subclassed form with all field
-        values fetched from the database. The result is cached in memory.
-        """
-        return self.get_specific()
-
-    @cached_property
-    def specific_deferred(self):
-        """
-        Returns this page in its most specific subclassed form without any
-        additional field values being fetched from the database. The result
-        is cached in memory.
-        """
-        return self.get_specific(deferred=True)
-
-    @cached_property
-    def specific_class(self):
-        """
-        Return the class that this page would be if instantiated in its
-        most specific form.
-
-        If the model class can no longer be found in the codebase, and the
-        relevant ``ContentType`` has been removed by a database migration,
-        the return value will be ``None``.
-
-        If the model class can no longer be found in the codebase, but the
-        relevant ``ContentType`` is still present in the database (usually a
-        result of switching between git branches without running or reverting
-        database migrations beforehand), the return value will be ``None``.
-        """
-        return self.cached_content_type.model_class()
-
-    @property
-    def cached_content_type(self):
-        """
-        Return this page's ``content_type`` value from the ``ContentType``
-        model's cached manager, which will avoid a database query if the
-        object is already in memory.
-        """
-        return ContentType.objects.get_for_id(self.content_type_id)
 
     @property
     def page_type_display_name(self):
@@ -3569,7 +3454,7 @@ class TaskManager(models.Manager):
         return self.filter(active=True)
 
 
-class Task(models.Model):
+class Task(SpecificMixin, models.Model):
     name = models.CharField(max_length=255, verbose_name=_("name"))
     content_type = models.ForeignKey(
         ContentType,
@@ -3620,28 +3505,6 @@ class Task(models.Model):
         # This is similar to doing cls._meta.verbose_name.title()
         # except this doesn't convert any characters to lowercase
         return capfirst(cls._meta.verbose_name)
-
-    @cached_property
-    def specific(self):
-        """
-        Return this Task in its most specific subclassed form.
-        """
-        # the ContentType.objects manager keeps a cache, so this should potentially
-        # avoid a database lookup over doing self.content_type. I think.
-        content_type = ContentType.objects.get_for_id(self.content_type_id)
-        model_class = content_type.model_class()
-        if model_class is None:
-            # Cannot locate a model class for this content type. This might happen
-            # if the codebase and database are out of sync (e.g. the model exists
-            # on a different git branch and we haven't rolled back migrations before
-            # switching branches); if so, the best we can do is return the task
-            # unchanged.
-            return self
-        elif isinstance(self, model_class):
-            # self is already the an instance of the most specific class
-            return self
-        else:
-            return content_type.get_object_for_this_type(id=self.id)
 
     task_state_class = None
 
@@ -4370,7 +4233,7 @@ class TaskStateManager(models.Manager):
             )
 
 
-class TaskState(models.Model):
+class TaskState(SpecificMixin, models.Model):
     """Tracks the status of a given Task for a particular revision."""
 
     STATUS_IN_PROGRESS = "in_progress"
@@ -4450,28 +4313,6 @@ class TaskState(models.Model):
             "revision_info": self.revision,
             "status": self.status,
         }
-
-    @cached_property
-    def specific(self):
-        """
-        Return this TaskState in its most specific subclassed form.
-        """
-        # the ContentType.objects manager keeps a cache, so this should potentially
-        # avoid a database lookup over doing self.content_type. I think.
-        content_type = ContentType.objects.get_for_id(self.content_type_id)
-        model_class = content_type.model_class()
-        if model_class is None:
-            # Cannot locate a model class for this content type. This might happen
-            # if the codebase and database are out of sync (e.g. the model exists
-            # on a different git branch and we haven't rolled back migrations before
-            # switching branches); if so, the best we can do is return the task state
-            # unchanged.
-            return self
-        elif isinstance(self, model_class):
-            # self is already the an instance of the most specific class
-            return self
-        else:
-            return content_type.get_object_for_this_type(id=self.id)
 
     @transaction.atomic
     def approve(self, user=None, update=True, comment=""):
